pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-rust"]
build-backend = "setuptools.build_meta"

[project]
name = "rustputo"
version = "0.1.3"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
    "pytools@git+https://github.com/willwiz/pytools@v1.1",
    "matplotlib",
    "numpy",
    "scipy-stubs[scipy]>=1.17.0.2",
]
license-files = ["LICENSE"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/test_*.py" = ["S101"]

[tool.setuptools.packages]
find = { where = ["src"] }
[[tool.setuptools-rust.ext-modules]]
target = "rustputo.rust.models"
path = "rust/models/Cargo.toml"
binding = "PyO3"


[dependency-groups]
dev = ["pytest>=9.0.2"]
